refactor bulk API code into separate file for maintainability
